integrate new PoS contract into genesis block; make stake type Long instead of Int; update extraction of bonds from tuplespace using query against PoS contract

+++ casper/src/main/scala/coop/rchain/casper/Estimator.scala
@@ -18,7 +18,7 @@ object Estimator {
   type BlockHash = ByteString
   type Validator = ByteString
 
-  implicit val decreasingOrder = Ordering[Int].reverse
+  implicit val decreasingOrder = Ordering[Long].reverse
 
   /**
     * When the BlockDag has an empty latestMessages, tips will return IndexedSeq(genesis)
@@ -28,11 +28,11 @@ object Estimator {
     @tailrec
     def sortChildren(blocks: IndexedSeq[BlockHash],
                      childMap: Map[BlockHash, Set[BlockHash]],
-                     scores: Map[BlockHash, Int]): IndexedSeq[BlockHash] = {
+                     scores: Map[BlockHash, Long]): IndexedSeq[BlockHash] = {
       // TODO: This ListContrib.sortBy will be improved on Thursday with Pawels help
       val newBlocks =
         ListContrib
-          .sortBy[BlockHash, Int](
+          .sortBy[BlockHash, Long](
             blocks.flatMap(replaceBlockHashWithChildren(childMap, _, scores)).distinct,
             scores)
       if (stillSame(blocks, newBlocks)) {
@@ -49,7 +49,7 @@ object Estimator {
       */
     def replaceBlockHashWithChildren(childMap: Map[BlockHash, Set[BlockHash]],
                                      b: BlockHash,
-                                     scores: Map[BlockHash, Int]): IndexedSeq[BlockHash] = {
+                                     scores: Map[BlockHash, Long]): IndexedSeq[BlockHash] = {
       val c: Set[BlockHash] = childMap.getOrElse(b, Set.empty[BlockHash]).filter(scores.contains)
       if (c.nonEmpty) {
         c.toIndexedSeq
@@ -70,15 +70,15 @@ object Estimator {
   }
 
   // TODO: Fix to stop at genesis/LFB
-  def buildScoresMap[F[_]: Monad: BlockStore](blockDag: BlockDag): F[Map[BlockHash, Int]] = {
+  def buildScoresMap[F[_]: Monad: BlockStore](blockDag: BlockDag): F[Map[BlockHash, Long]] = {
     def hashParents(hash: BlockHash): F[List[BlockHash]] =
       for {
         b <- unsafeGetBlock[F](hash)
       } yield parentHashes(b).toList
 
-    def addValidatorWeightDownSupportingChain(scoreMap: Map[BlockHash, Int],
+    def addValidatorWeightDownSupportingChain(scoreMap: Map[BlockHash, Long],
                                               validator: Validator,
-                                              latestBlockHash: BlockHash): F[Map[BlockHash, Int]] =
+                                              latestBlockHash: BlockHash): F[Map[BlockHash, Long]] =
       for {
         updatedScoreMap <- DagOperations
                             .bfTraverseF[F, BlockHash](List(latestBlockHash))(hashParents)
@@ -86,7 +86,7 @@ object Estimator {
                               case (acc, hash) =>
                                 for {
                                   b               <- unsafeGetBlock[F](hash)
-                                  currScore       = acc.getOrElse(hash, 0)
+                                  currScore       = acc.getOrElse(hash, 0L)
                                   validatorWeight <- weightFromValidator[F](b, validator)
                                 } yield acc.updated(hash, currScore + validatorWeight)
                             }
@@ -98,7 +98,7 @@ object Estimator {
       *
       * TODO: Add test where this matters
       */
-    def addValidatorWeightToImplicitlySupported(scoreMap: Map[BlockHash, Int],
+    def addValidatorWeightToImplicitlySupported(scoreMap: Map[BlockHash, Long],
                                                 childMap: Map[BlockHash, Set[BlockHash]],
                                                 validator: Validator,
                                                 latestBlockHash: BlockHash) =
@@ -112,8 +112,8 @@ object Estimator {
                 for {
                   c <- ProtoUtil.unsafeGetBlock[F](cHash)
                   result = if (ProtoUtil.parentHashes(c).size > 1 && c.sender != validator) {
-                    val currScore       = acc2.getOrElse(cHash, 0)
-                    val validatorWeight = ProtoUtil.weightMap(c).getOrElse(validator, 0)
+                    val currScore       = acc2.getOrElse(cHash, 0L)
+                    val validatorWeight = ProtoUtil.weightMap(c).getOrElse(validator, 0L)
                     acc2.updated(cHash, currScore + validatorWeight)
                   } else {
                     acc2
@@ -123,20 +123,21 @@ object Estimator {
         }
 
     for {
-      scoresMap <- Foldable[List].foldM(blockDag.latestMessages.toList, Map.empty[BlockHash, Int]) {
-                    case (acc, (validator: Validator, latestBlock: BlockMessage)) =>
-                      for {
-                        postValidatorWeightScoreMap <- addValidatorWeightDownSupportingChain(
-                                                        acc,
-                                                        validator,
-                                                        latestBlock.blockHash)
-                        postImplicitlySupportedScoreMap <- addValidatorWeightToImplicitlySupported(
-                                                            postValidatorWeightScoreMap,
-                                                            blockDag.childMap,
-                                                            validator,
-                                                            latestBlock.blockHash)
-                      } yield postImplicitlySupportedScoreMap
-                  }
+      scoresMap <- Foldable[List]
+                    .foldM(blockDag.latestMessages.toList, Map.empty[BlockHash, Long]) {
+                      case (acc, (validator: Validator, latestBlock: BlockMessage)) =>
+                        for {
+                          postValidatorWeightScoreMap <- addValidatorWeightDownSupportingChain(
+                                                          acc,
+                                                          validator,
+                                                          latestBlock.blockHash)
+                          postImplicitlySupportedScoreMap <- addValidatorWeightToImplicitlySupported(
+                                                              postValidatorWeightScoreMap,
+                                                              blockDag.childMap,
+                                                              validator,
+                                                              latestBlock.blockHash)
+                        } yield postImplicitlySupportedScoreMap
+                    }
     } yield scoresMap
   }
 }
