+++ casper/src/test/scala/coop/rchain/casper/util/comm/BlockApproverProtocolTest.scala
@@ -20,7 +20,7 @@ class BlockApproverProtocolTest extends FlatSpec with Matchers {
   "BlockApproverProtocol" should "respond to valid ApprovedBlockCandidates" in {
     val n                          = 8
     val (validatorSk, validatorPk) = Ed25519.newKeyPair
-    val bonds                      = Map(validatorPk -> 10)
+    val bonds                      = Map(validatorPk -> 10L)
     val (approver, node)           = createProtocol(n, Seq.empty, validatorSk, bonds)
     val unapproved                 = createUnapproved(n, node.genesis)
     import node._
@@ -36,7 +36,7 @@ class BlockApproverProtocolTest extends FlatSpec with Matchers {
   it should "log a warning for invalid ApprovedBlockCandidates" in {
     val n                          = 8
     val (validatorSk, validatorPk) = Ed25519.newKeyPair
-    val bonds                      = Map(validatorPk -> 10)
+    val bonds                      = Map(validatorPk -> 10L)
     val (approver, node)           = createProtocol(n, Seq.empty, validatorSk, bonds)
     val differentUnapproved1       = createUnapproved(n / 2, node.genesis) //wrong number of signatures
     val differentUnapproved2       = createUnapproved(n, BlockMessage.defaultInstance) //wrong block
@@ -62,7 +62,7 @@ object BlockApproverProtocolTest {
       requiredSigs: Int,
       wallets: Seq[PreWallet],
       sk: Array[Byte],
-      bonds: Map[Array[Byte], Int]): (BlockApproverProtocol, HashSetCasperTestNode[Id]) = {
+      bonds: Map[Array[Byte], Long]): (BlockApproverProtocol, HashSetCasperTestNode[Id]) = {
     import monix.execution.Scheduler.Implicits.global
 
     val runtimeDir     = BlockStoreTestFixture.dbDir
